Finalise pandification
